test(unit): fix failing tests
