+++ app/process/handler/fee-message/handler_test.go
@@ -24,7 +24,6 @@ import (
 	"github.com/limechain/hedera-eth-bridge-validator/constants"
 	"github.com/limechain/hedera-eth-bridge-validator/test/mocks"
 	"github.com/limechain/hedera-eth-bridge-validator/test/mocks/service"
-	"github.com/stretchr/testify/mock"
 	"testing"
 )
 
@@ -59,7 +58,7 @@ func Test_Handle(t *testing.T) {
 	mockedService.On("InitiateNewTransfer", mt).Return(tx, nil)
 	mockedService.On("ProcessNativeTransfer", mt).Return(nil)
 
-	ctHandler.Handle(mt)
+	ctHandler.Handle(&mt)
 
 	mockedService.AssertCalled(t, "InitiateNewTransfer", mt)
 	mockedService.AssertCalled(t, "ProcessNativeTransfer", mt)
@@ -115,24 +114,6 @@ func Test_Handle_ProcessNativeTransfer_Fails(t *testing.T) {
 
 	mockedService.On("InitiateNewTransfer", mt).Return(tx, nil)
 	mockedService.On("ProcessNativeTransfer", mt).Return(errors.New("some-error"))
-	mockedService.AssertNotCalled(t, "ProcessWrappedTransfer", mock.Anything)
-
-	ctHandler.Handle(&mt)
-}
-
-func Test_Handle_ProcessWrappedTransfer_Fails(t *testing.T) {
-	ctHandler, mockedService := InitializeHandler()
-
-	tx := &entity.Transfer{
-		TransactionID: mt.TransactionId,
-		Receiver:      mt.Receiver,
-		Amount:        mt.Amount,
-		Status:        transfer.StatusInitial,
-	}
-
-	mockedService.On("InitiateNewTransfer", mt).Return(tx, nil)
-	mockedService.On("ProcessWrappedTransfer", mt).Return(errors.New("some-error"))
-	mockedService.AssertNotCalled(t, "ProcessNativeTransfer", mock.Anything)
 
 	ctHandler.Handle(&mt)
 }
