+++ app/process/watcher/transfer/watcher.go
@@ -162,11 +162,11 @@ func (ctw Watcher) processTransaction(tx mirror_node.Transaction, q qi.Queue) {
 	targetChainAsset := ctw.mappings.NativeToWrapped(asset, 0, targetChainId)
 	if targetChainAsset == "" {
 		nativeAsset = ctw.mappings.WrappedToNative(asset, 0)
-		targetChainAsset = nativeAsset.Asset
 		if nativeAsset == nil {
 			ctw.logger.Errorf("[%s] - Could not parse asset [%s] to its target chain correlation", tx.TransactionID, asset)
 			return
 		}
+		targetChainAsset = nativeAsset.Asset
 		if nativeAsset.ChainId != targetChainId {
 			ctw.logger.Errorf("[%s] - Wrapped to Wrapped transfers currently not supported [%s] - [%d] for [%d]", tx.TransactionID, nativeAsset.Asset, nativeAsset.ChainId, targetChainId)
 			return
